requirements.txt
aiohttp==3.8.5
discord.py==2.3.2
PyYAML==6.0.1